Add category deletion with product reassignment, auto-generated product references, and performer full names
Deleting a category now nullifies the category on its products
(displayed as Uncategorized) instead of blocking the delete, wrapped
in one transaction. Category prefixes drive an auto-suggested next
product reference (GET /products/next-ref), and sale/stock-operation
responses now surface the performer's full name instead of username.

+++ backend/src/main/java/com/shopstock/entity/StockOperation.java
@@ -26,7 +26,7 @@ public class StockOperation {
 
     @NotNull
     @Column(nullable = false)
-    private Integer quantityChange;
+    private Float quantityChange;
 
     @Column(nullable = false, updatable = false)
     private LocalDateTime operationDate;
@@ -67,11 +67,11 @@ public class StockOperation {
         this.product = product;
     }
 
-    public Integer getQuantityChange() {
+    public Float getQuantityChange() {
         return quantityChange;
     }
 
-    public void setQuantityChange(Integer quantityChange) {
+    public void setQuantityChange(Float quantityChange) {
         this.quantityChange = quantityChange;
     }
